pyproject.toml
[project]
name = "doujia"
version = "0.1.0"
description = "Yuchen's personal beancount ledger project"
authors = [{ name = "e7h4n", email = "perfectworks@gmail.com" }]
dependencies = [
    "setuptools>=69.2.0",
    "Flask==3.0.2",
    "beancount==3.0.0",
    "beangrow@git+https://github.com/e7h4n/beangrow.git",
    "beanquery@git+https://github.com/beancount/beanquery.git",
    "fava==1.30",
    "frozendict==2.4.0",
    "logzero==1.7.0",
    "protobuf==3.20.3",
    "certifi>=2024.8.30",
    "passkeys>=1.1.3",
    "flask-cors>=5.0.0",
    "pytest-mock>=3.14.0",
    "Flask-APScheduler>=1.13.1"
]
requires-python = "~=3.12.0"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm.dev-dependencies]
lint = ["black", "flake8", "pre-commit", "pylint"]
test = ["coverage", "freezegun", "pyfakefs", "pytest", "pytest-timeout"]
dev = ["-e file:///${PROJECT_ROOT}/#egg=doujia"]

[tool.pdm]
distribution = false
